requirements.txt
streamlit
pandas
python-docx
openai
python-dotenv